refactor(重构部分功能): 重构部分功能

+++ page/page.go
@@ -3,20 +3,20 @@ package page
 import "math"
 
 type Page struct {
-	Page      int64 `json:"Page"`
-	PageSize  int64 `json:"PageSize"`
-	Total     int64 `json:"Total"`
-	PrevPage  int64 `json:"PrevPage"`
-	NextPage  int64 `json:"NextPage"`
-	TotalPage int64 `json:"TotalPage"`
-	Limit     int64 `json:"-"`
-	Offset    int64 `json:"-"`
+	Page      int `json:"Page"`
+	PageSize  int `json:"PageSize"`
+	Total     int `json:"Total"`
+	PrevPage  int `json:"PrevPage"`
+	NextPage  int `json:"NextPage"`
+	TotalPage int `json:"TotalPage"`
+	Limit     int `json:"-"`
+	Offset    int `json:"-"`
 }
 
-func Paginator(page, pageSize, total int64) *Page {
+func Paginator(page, pageSize, total int) *Page {
 
 	//根据nums总数，和prePage每页数量 生成分页总数
-	totalPage := int64(math.Ceil(float64(total) / float64(pageSize))) //page总数
+	totalPage := int(math.Ceil(float64(total) / float64(pageSize))) //page总数
 	if page > totalPage {
 		page = totalPage
 	}
